--- index.html
@@ -1,22 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" style="height: 100%">
-  <head>
-    <link rel="icon" type="image/x-icon" href="./src/assets/brain.svg">
-    <link rel="stylesheet" href="style.css">
-    <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Share+Tech+Mono">
-    <meta charset="utf-8">
-    <title>Galaxy Brain</title>
-  </head>
-  <body>
-    <script src="bundle.js"></script>
-        <input type="range" min="0" max="3" value="0" step="1" list="stages">
-        <datalist id="stages">
-          <option>0</option>
-          <option>1</option>
-          <option>2</option>
-          <option>3</option>
-        </datalist>
-        <canvas id="matrix-rain"></canvas>
-        <canvas id="galaxy-brain" class="stage-0"></canvas>
-  </body>
-</html>
