--- matrix-rain.d.ts
@@ -1,8 +1,0 @@
-declare module "@ojdom/matrix-rain" {
-    export class MatrixController {
-		constructor(options?: {}, canvas?: HTMLCanvasElement);
-		start: () => void;
-		disable: () => void;
-		stop: () => void;
-	}
-}
