+++ package.json
@@ -18,13 +18,16 @@
     "@types/jquery": "^3.5.4",
     "@typescript-eslint/eslint-plugin": "^4.27.0",
     "@typescript-eslint/parser": "^4.27.0",
+    "css-loader": "^6.8.1",
     "eslint": "^7.29.0",
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-config-airbnb-typescript": "^12.3.1",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-prettier": "^3.4.0",
+    "html-webpack-plugin": "^5.5.3",
     "prettier": "2.3.1",
+    "style-loader": "^3.3.3",
     "ts-loader": "^9.4.2",
     "ts-shader-loader": "^2.0.2",
     "typescript": "^4.1.2",
