+++ src/index.html
@@ -1,0 +1,18 @@
+<!DOCTYPE html>
+<html lang="en" style="height: 100%">
+  <head>
+    <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Share+Tech+Mono">
+    <meta charset="utf-8">
+  </head>
+  <body>
+    <input type="range" min="0" max="3" value="0" step="1" list="stages">
+    <datalist id="stages">
+      <option>0</option>
+      <option>1</option>
+      <option>2</option>
+      <option>3</option>
+    </datalist>
+    <canvas id="matrix-rain"></canvas>
+    <canvas id="galaxy-brain" class="stage-0"></canvas>
+  </body>
+</html>
